+++ log.txt
@@ -154,3 +154,8 @@
 {"level":"info","message":"trump command used by AirFusion Time: Thu Nov 25 2021 11:24:15 GMT-0800 (Pacific Standard Time) Guild: db8bottest"}
 {"level":"info","message":"trump command used by AirFusion Time: Thu Nov 25 2021 11:24:18 GMT-0800 (Pacific Standard Time) Guild: db8bottest"}
 {"level":"info","message":"trump command used by AirFusion Time: Thu Nov 25 2021 11:24:22 GMT-0800 (Pacific Standard Time) Guild: db8bottest"}
+{"level":"info","message":"get command used by AirFusion Time: Thu Nov 25 2021 17:19:53 GMT-0800 (Pacific Standard Time) Guild: db8bottest"}
+{"level":"info","message":"get command used by AirFusion Time: Thu Nov 25 2021 17:20:53 GMT-0800 (Pacific Standard Time) Guild: db8bottest"}
+{"level":"info","message":"get command used by AirFusion Time: Thu Nov 25 2021 17:21:25 GMT-0800 (Pacific Standard Time) Guild: db8bottest"}
+{"level":"info","message":"get command used by AirFusion Time: Thu Nov 25 2021 17:22:30 GMT-0800 (Pacific Standard Time) Guild: db8bottest"}
+{"level":"info","message":"get command used by AirFusion Time: Thu Nov 25 2021 17:25:36 GMT-0800 (Pacific Standard Time) Guild: db8bottest"}
