+++ package.json
@@ -37,6 +37,8 @@
     "discord.js": "^13.1.0",
     "dotenv": "^10.0.0",
     "ipfs": "^0.58.4",
+    "jsdom": "^18.1.1",
+    "mhtml2html": "^3.0.0",
     "mongodb": "^4.1.1",
     "ms": "^2.1.3",
     "pngjs": "^6.0.0",
